requirements.txt
psycopg2~=2.9.9
requests~=2.31.0
aiogram~=3.4.1
aiohttp~=3.9.3
python-dotenv~=1.0.1
Django~=5.0.6
djangorestframework~=3.15.1